+++ test/call-step-from-step/test.mjs
@@ -1,0 +1,21 @@
+import {
+  test,
+  TestDir,
+  execPlaywrightTest,
+  execPlaywrightTestWithError,
+  DEFAULT_CMD,
+} from '../helpers.mjs';
+
+const testDir = new TestDir(import.meta);
+
+test(`${testDir.name} (success)`, () => {
+  execPlaywrightTest(testDir.name, `${DEFAULT_CMD} --project success`);
+});
+
+test(`${testDir.name} (invalid invocation)`, () => {
+  execPlaywrightTestWithError(
+    testDir.name,
+    ['Missings fixtures: todos, $testInfo'],
+    `${DEFAULT_CMD} --project fail`,
+  );
+});