+++ test/bdd-syntax-cucumber-style/test.mjs
@@ -1,3 +1,0 @@
-import { test, getTestName, execPlaywrightTest } from '../helpers.mjs';
-
-test(getTestName(import.meta), (t) => execPlaywrightTest(t.name));