+++ test/bdd-syntax/steps-pw-style/steps.ts
@@ -80,3 +80,5 @@ Step('Tags are {string}', async ({ $tags, tagsFromCustomFixture }, tags: string)
   expect($tags.join(' ')).toEqual(tags);
   expect(tagsFromCustomFixture.join(' ')).toEqual(tags);
 });
+
+Step('Skipped step', () => {});