+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## dev
+## 6.1.0
 * feature: add new special tags ([#123](https://github.com/vitalets/playwright-bdd/issues/123), [#126](https://github.com/vitalets/playwright-bdd/issues/126))
 * feature: support new Playwright tags ([#98](https://github.com/vitalets/playwright-bdd/issues/98))
 * feature: add option `--unused-steps` to `bddgen export` ([#113](https://github.com/vitalets/playwright-bdd/issues/113))